Add `processPendingChanges()` method to persistent store connector
This enables subclass implementations to notify us to immediately process the batch as soon as all their data has been fetched. This method is not required, however it allows us to cut out a few extra milliseconds of unneccessary delay in waiting for the throttling interval to trigger. In other words when we know we’ve received all our data for the current fetch, we can call this method.