Remove dependence on NSObject, NSSortDescriptor, and NSPredicate
Forcing result objects to conform to NSObject is limiting. One area where this was limiting was the fact that Swift structs could not be used as result objects.

The reason result objects needed to conform to NSObject was simply in order for the library to use key-value coding to work with `sectionNameKeyPath` as well as the sort descriptors and predicate configured on the store request object.

This commit modernizes the whole system by using the same standard closures used in the Swift `filter` and `sort` methods, and makes the results controller significantly more flexible by allowing struct based model objects to be used with it.