Fix, simplify, and cleanup problems with batching logic
I was observing some problems with the way the batch controller was handling results and missing results objects or changes.

Specifically the problems seemed to manifest if the `processesChangesImmediately` flag was set to true, or the `processPendingChanges()` method used (they effectively do the same things).

Focuses on only using the throttler to throttle calls to the `flush()` method rather then trying to put everything on the throttlers background queue which was just making things difficult to track and follow anyways.

The `isBatching` flag is now also a computer property rather then manually set.