Cache section names (i.e. sectionKeyValue)
In the example project this fixes an issue where deleted core data objects are turned into faults after saving the context. This was causing a crash since accessing the properties of a fault returns “nil” as opossed to the actual expected value.

Caching the section names resolves the above issue (which is a CoreData specific use case), however this also has the added benefit of being more efficient.