Make CRUDStore the BaseStore
The most fundemental aspect of a store is the type/model it interacts with (not the request type). It’s entirely possible to have 2 stores that interact with the same model type/model but fetch from different sources and therefore potentially different request types.

This architectures provides more flexibility when subclassing or referencing stores generically.