Update `InstanceIdentifiable` protocol to support all Hashable ID types
The `InstanceIdentifiable` protocol was created as a backwards compatible version of Swift’s `Identifiable` protocol. Making the `ID` type Hashable rather than forcing it to be a String, enables true backwards compatibility.

This protocol can be removed and replaced with `Identifiable` when the minimum deployment target is updated to iOS 13.